renamed Popup constants file, modified types for reducer , ordered alphabetically messages.json
