+++ src/app/content/highlights/components/SummaryPopup/ChapterFilter.tsx
@@ -6,7 +6,7 @@ import Checkbox from '../../../../components/Checkbox';
 import { textStyle } from '../../../../components/Typography/base';
 import { match, not } from '../../../../fpUtils';
 import theme from '../../../../theme';
-import { filters } from '../../../styles/constants';
+import { filters } from '../../../styles/PopupConstants';
 import { setSummaryFilters } from '../../actions';
 import { highlightLocationFilters, highlightLocationFiltersWithContent, summaryLocationFilters } from '../../selectors';
 import ColorIndicator from '../ColorIndicator';
