+++ src/app/content/highlights/components/SummaryPopup/ColorFilter.tsx
@@ -8,7 +8,7 @@ import Checkbox from '../../../../components/Checkbox';
 import { textStyle } from '../../../../components/Typography/base';
 import { match, not } from '../../../../fpUtils';
 import theme from '../../../../theme';
-import { filters } from '../../../styles/constants';
+import { filters } from '../../../styles/PopupConstants';
 import { setSummaryFilters } from '../../actions';
 import { highlightStyles } from '../../constants';
 import { highlightColorFiltersWithContent, summaryColorFilters } from '../../selectors';
