+++ src/app/content/studyGuides/types.ts
@@ -2,8 +2,6 @@ import { HighlightsSummary } from '@openstax/highlighter/dist/api';
 
 export interface State {
   isEnabled: boolean;
-  summary: {
-    open: boolean,
-    studyGuides: HighlightsSummary | null,
-  };
+  open: boolean;
+  summary: HighlightsSummary | null;
 }
