+++ src/app/content/styles/PopupConstants.ts
